move loadResponses to /services
